+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added some config descriptions.
+
+### Fixed
+
+- Fixed errors in calculating reward and shop item quantities.
+
 ## [1.1.0] - 2024-06-07
 
 ### Added